ai-app/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    ./scripts/upgrade.sh
#
alembic==1.13.1
altair==5.2.0
annotated-types==0.6.0
anyio==4.2.0
attrs==23.2.0
beautifulsoup4==4.12.3
blinker==1.7.0
boto3==1.34.34
botocore==1.34.34
cachetools==5.3.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
distro==1.9.0
docker==7.0.0
fastapi==0.109.1
gitdb==4.0.11
gitpython==3.1.41
h11==0.14.0
httpcore==1.0.2
httpx==0.27.0
idna==3.6
importlib-metadata==7.0.1
jinja2==3.1.3
jmespath==1.0.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
mako==1.3.2
markdown-it-py==3.0.0
markupsafe==2.1.5
mdurl==0.1.2
mypy==1.8.0
mypy-extensions==1.0.0
numpy==1.26.3
ollama==0.2.0
openai==1.11.0
packaging==23.2
pandas==2.2.0
pgvector==0.2.4
phidata==2.3.45
pillow==10.2.0
protobuf==4.25.2
psycopg[binary]==3.1.17
psycopg-binary==3.1.17
pyarrow==15.0.0
pydantic==2.6.0
pydantic-core==2.16.1
pydantic-settings==2.1.0
pydeck==0.8.1b0
pygments==2.17.2
pypdf==4.0.1
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.1
pyyaml==6.0.1
referencing==0.33.0
regex==2023.12.25
requests==2.31.0
rich==13.7.0
rpds-py==0.17.1
ruff==0.2.0
s3transfer==0.10.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
sqlalchemy==2.0.25
starlette==0.35.1
streamlit==1.31.0
tenacity==8.2.3
tiktoken==0.5.2
toml==0.10.2
tomli==2.0.1
toolz==0.12.1
tornado==6.4
tqdm==4.66.1
typer==0.9.0
types-beautifulsoup4==4.12.0.20240106
types-html5lib==1.1.11.20240106
types-pillow==10.2.0.20240125
typing-extensions==4.9.0
tzdata==2023.4
tzlocal==5.2
urllib3==2.0.7
uvicorn==0.27.0.post1
validators==0.22.0
zipp==3.17.0
